Bump `@npmcli/arborist` and `pacote`
Upgrade these dependencies to newer versions in order to resolve 13
deprecation warnings (some are still present due to the dependency on
`get-dep-tree`, which would need to receive similar upgrades). This
comes at the cost of dropping support for Node.js `^14.18` and `=16.13`.
These are the minimum (major version) upgrades necessary to achieve this
goal.